inertia-backend/requirements.txt
fastapi>=0.111.0,<1.0.0
uvicorn[standard]>=0.29.0,<1.0.0
google-genai>=1.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
redis>=5.0.4,<6.0.0
httpx>=0.27.0,<1.0.0
pytest>=8.2.0,<9.0.0
